refactor(apply): (AHWR-2044) extract shared apply helpers to neutral module #patch
